Move service worker registration code to a subdirectory
